Make different class to isolate the scrren draw mechanism
